Fix unduly harsh quiz result messages
The mobile apps have been showing users the quiz result message that
corresponds to `score - 1`. For instance, on this quiz:

https://www.theguardian.com/uk/quiz/2013/jan/27/british-citizenship-test-quiz-new
https://content.guardianapis.com/atom/quiz/5eeb9291-80e2-41a9-8d9d-603cedd209f6

...if you complete the quiz with a score of 8 out of 10 answers, you get
the correct score displayed ("You scored: 8/10"), but the message shown is
the message for 7/10, which includes the phrase "...on this 10-question
test you've come in at 70%..." and exposes the problem.

This off-by-one error has been present since the functionality to show the
result text was added with commit 102c32bfc, but obviously only affects
quizzes displaying on the mobile apps (iOS and Android), not the website,
though that did have https://github.com/guardian/frontend/pull/15276 last
year :)

Thanks to avid Guardian reader Michael Goldrei for reporting this issue
with a good screenshot and url!